Revert "Fixed activity build."
This reverts commit eef09c8e23ce8a2d7f567985f99ed92a3b8d3251.
