--- WcfServiceLibrary1/WcfServiceLibrary1.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WcfServiceLibrary1", "WcfServiceLibrary1\WcfServiceLibrary1.csproj", "{A26C0BD9-E8F7-4221-95E2-C864F7F5B662}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A26C0BD9-E8F7-4221-95E2-C864F7F5B662}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A26C0BD9-E8F7-4221-95E2-C864F7F5B662}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A26C0BD9-E8F7-4221-95E2-C864F7F5B662}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A26C0BD9-E8F7-4221-95E2-C864F7F5B662}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
