--- WcfServiceLibrary1/WcfServiceLibrary1/App.config
@@ -1,50 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-
-  <system.web>
-    <compilation debug="true" />
-  </system.web>
-  <!-- When deploying the service library project, the content of the config file must be added to the host's 
-  app.config file. System.Configuration does not support config files for libraries. -->
-  <system.serviceModel>
-    <services>
-      <service name="WcfServiceLibrary1.Service1">
-        <host>
-          <baseAddresses>
-            <add baseAddress = "http://localhost:8732/Design_Time_Addresses/WcfServiceLibrary1/Service1/" />
-          </baseAddresses>
-        </host>
-        <!-- Service Endpoints -->
-        <!-- Unless fully qualified, address is relative to base address supplied above -->
-        <endpoint address ="" binding="wsHttpBinding" contract="WcfServiceLibrary1.IService1">
-          <!-- 
-              Upon deployment, the following identity element should be removed or replaced to reflect the 
-              identity under which the deployed service runs.  If removed, WCF will infer an appropriate identity 
-              automatically.
-          -->
-          <identity>
-            <dns value="localhost"/>
-          </identity>
-        </endpoint>
-        <!-- Metadata Endpoints -->
-        <!-- The Metadata Exchange endpoint is used by the service to describe itself to clients. --> 
-        <!-- This endpoint does not use a secure binding and should be secured or removed before deployment -->
-        <endpoint address="mex" binding="mexHttpBinding" contract="IMetadataExchange"/>
-      </service>
-    </services>
-    <behaviors>
-      <serviceBehaviors>
-        <behavior>
-          <!-- To avoid disclosing metadata information, 
-          set the value below to false and remove the metadata endpoint above before deployment -->
-          <serviceMetadata httpGetEnabled="True"/>
-          <!-- To receive exception details in faults for debugging purposes, 
-          set the value below to true.  Set to false before deployment 
-          to avoid disclosing exception information -->
-          <serviceDebug includeExceptionDetailInFaults="False" />
-        </behavior>
-      </serviceBehaviors>
-    </behaviors>
-  </system.serviceModel>
-
-</configuration>
