+++ wcf_service/README.md
@@ -1,0 +1,20 @@
+# WCF Tutorial
+
+##Learning Competencies
+
+* Learn how WCFs work and how to build one yourself
+
+##Summary
+
+Microsoft has moved to consolidate all of their web service frameworks into their new WCF (Windows Communication Foundation) projects. In this tutorial you will learn to build a web service using it in Visual Studio
+
+##Instructions:
+
+1. Goto the WCF tutorial [here](http://www.tutorialspoint.com/wcf/wcf_creating_service.htm) to run through setting up the basic project. Follow the instructions to convert the base app to be a service that does basic calculator functions.
+
+2. Once you have your app running and have ran through all the instructions, commit your work and move to the next step.
+
+3. Next, add functionality to the app to convert celcius to farhenheit and the other way. You have creative control to make this one or two different endpoints to hit, whichever you see fit will do.
+
+4. Submit your work through a pull request with the twilio work
+
