+++ wcf_service/SyndicationServiceLibrary1/SyndicationServiceLibrary1.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SyndicationServiceLibrary1", "SyndicationServiceLibrary1\SyndicationServiceLibrary1.csproj", "{60EA77ED-F8F1-4CDE-A2E9-F41E35F34A2C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{60EA77ED-F8F1-4CDE-A2E9-F41E35F34A2C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{60EA77ED-F8F1-4CDE-A2E9-F41E35F34A2C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{60EA77ED-F8F1-4CDE-A2E9-F41E35F34A2C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{60EA77ED-F8F1-4CDE-A2E9-F41E35F34A2C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
