+++ wcf_service/SyndicationServiceLibrary1/SyndicationServiceLibrary1/App.config
@@ -1,0 +1,25 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <system.web>
+    <compilation debug="true" />
+  </system.web>
+  <system.serviceModel>
+    <services>
+      <service name="SyndicationServiceLibrary1.Feed1">
+        <host>
+          <baseAddresses>
+            <add baseAddress="http://localhost:8732/Design_Time_Addresses/SyndicationServiceLibrary1/" />
+          </baseAddresses>
+        </host>
+        <endpoint contract="SyndicationServiceLibrary1.IFeed1" address="Feed1" binding="webHttpBinding" behaviorConfiguration="SyndicationServiceLibrary1.Feed1Behavior"/>
+      </service>
+    </services>
+    <behaviors>
+      <endpointBehaviors>
+        <behavior name="SyndicationServiceLibrary1.Feed1Behavior">
+          <webHttp/>
+        </behavior>
+      </endpointBehaviors>
+    </behaviors>
+  </system.serviceModel>
+</configuration>
