+++ wcf_service/SyndicationServiceLibrary1/SyndicationServiceLibrary1/SyndicationServiceLibrary1.csproj.user
